+++ token/ts/src/instructions/revoke.ts
@@ -2,7 +2,7 @@ import { struct, u8 } from '@solana/buffer-layout';
 import { PublicKey, Signer, TransactionInstruction } from '@solana/web3.js';
 import { TOKEN_PROGRAM_ID } from '../constants';
 import { TokenInstruction } from './types';
-import { addSigners } from './utils';
+import { addSigners } from './internal';
 
 const dataLayout = struct<{ instruction: TokenInstruction }>([u8('instruction')]);
